requirements.txt
keyboard
pyperclip
customtkinter
groq
pystray
Pillow
pyinstaller
